Create EntityDeleter to have small class with less dependecies
